Main-API/requirements.txt
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mysql-connector-python==8.0.27
protobuf==3.19.1
pycparser==2.21
PyJWT==2.3.0
PyYAML==6.0
six==1.16.0
vboxapi==1.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
